// src/frontend/src/declarations/backend.did.d.ts
/* eslint-disable */

// @ts-nocheck

// This file was automatically generated by @caffeinelabs/bindgen@0.1.0.
// You should NOT make any changes in this file as it will be overwritten.
// Additionally, you should also exclude this file from your linter and/or formatter to prevent it from being checked or modified.

import type { ActorMethod } from '@icp-sdk/core/agent';
import type { IDL } from '@icp-sdk/core/candid';
import type { Principal } from '@icp-sdk/core/principal';

export type Time = bigint;
export interface Transaction {
  'id' : string,
  'transactionType' : TransactionType,
  'user' : Principal,
  'timestamp' : Time,
  'amount' : bigint,
}
export type TransactionType = { 'taskCompletion' : null } |
  { 'checkIn' : null } |
  { 'adWatch' : null } |
  { 'spin' : null } |
  { 'referralBonus' : null };
export interface UserProfile {
  'referralCode' : string,
  'coins' : bigint,
  'totalEarned' : bigint,
  'checkInStreak' : bigint,
  'referredBy' : [] | [string],
  'lastCheckIn' : [] | [Time],
  'lastSpin' : [] | [Time],
  'adsWatchedToday' : bigint,
}
export type UserRole = { 'admin' : null } |
  { 'user' : null } |
  { 'guest' : null };
export interface _SERVICE {
  '_initializeAccessControlWithSecret' : ActorMethod<[string], undefined>,
  'approveWithdrawal' : ActorMethod<[string], undefined>,
  'assignCallerUserRole' : ActorMethod<[Principal, UserRole], undefined>,
  'dailyCheckIn' : ActorMethod<[], bigint>,
  'getCallerUserProfile' : ActorMethod<[], UserProfile>,
  'getCallerUserRole' : ActorMethod<[], UserRole>,
  'getTopUsers' : ActorMethod<[bigint], Array<[Principal, UserProfile]>>,
  'getUserProfile' : ActorMethod<[Principal], UserProfile>,
  'getUserTransactions' : ActorMethod<[Principal], Array<Transaction>>,
  'isCallerAdmin' : ActorMethod<[], boolean>,
  'registerUser' : ActorMethod<[[] | [string]], undefined>,
  'rejectWithdrawal' : ActorMethod<[string], undefined>,
  'saveCallerUserProfile' : ActorMethod<[UserProfile], undefined>,
  'spinWheel' : ActorMethod<[], bigint>,
  'submitWithdrawalRequest' : ActorMethod<[bigint, string], undefined>,
  'watchAd' : ActorMethod<[], bigint>,
}
export declare const idlService: IDL.ServiceClass;
export declare const idlInitArgs: IDL.Type[];
export declare const idlFactory: IDL.InterfaceFactory;
export declare const init: (args: { IDL: typeof IDL }) => IDL.Type[];